requirements.txt
fastmcp
langgraph
aws-cdk-lib
constructs
boto3